refactor: remove test code from production MCP layer (SPI-612)
- Move TestResourceProvider from production to test fixtures
- Move Application.configureForTesting() to test utilities
- Remove unused MCPConfiguration.forTesting() method
- Update 6 test files with corrected imports
- Clean separation: zero test code in production JAR

Production benefits:
- Reduced JAR size (no test fixtures bundled)
- Better architecture (clear separation of concerns)
- Improved security (no test code in production)

Fixes SPI-612: Complete architectural separation achieved
