+++ src/main/kotlin/io/spiralhouse/cycletime/domain/services/TimeProvider.kt
@@ -10,15 +10,3 @@ interface TimeProvider {
 class SystemTimeProvider : TimeProvider {
     override fun now(): Instant = Clock.System.now()
 }
-
-class MockTimeProvider(private var currentTime: Instant = Clock.System.now()) : TimeProvider {
-    override fun now(): Instant = currentTime
-
-    fun setTime(time: Instant) {
-        currentTime = time
-    }
-
-    fun advance(duration: kotlin.time.Duration) {
-        currentTime = currentTime + duration
-    }
-}
