+++ src/test/kotlin/io/spiralhouse/cycletime/api/HealthEndpointTest.kt
@@ -13,7 +13,7 @@ import io.ktor.server.response.*
 import io.ktor.server.application.*
 import io.ktor.server.plugins.di.*
 import io.spiralhouse.cycletime.module
-import io.spiralhouse.cycletime.configureForTesting
+import io.spiralhouse.cycletime.infrastructure.di.modules.test.configureForTesting
 import io.spiralhouse.cycletime.infrastructure.database.TestDatabaseFactory
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.jsonObject
